requirements.txt
flask
requests
websocket-client==1.3.3
pymysql
